factor out more MMAv5 pipelining methods
