+++ lib/Dialect/TritonGPU/Transforms/Pipeliner/TC05MMAPipeline.cpp
@@ -17,42 +17,7 @@ namespace tt = mlir::triton;
 namespace ttg = mlir::triton::gpu;
 namespace ttng = mlir::triton::nvidia_gpu;
 
-//===----------------------------------------------------------------------===//
-// Utilities
-//===----------------------------------------------------------------------===//
-
-std::optional<std::pair<ttng::TMEMAllocOp, ttng::TMEMLoadOp>>
-ttng::getTMemAllocAndLoad(ttng::MMAv5OpInterface mmaOp) {
-  auto acc = mmaOp->getOperand(2).getDefiningOp<ttng::TMEMAllocOp>();
-  if (!acc || acc->getParentRegion() != mmaOp->getParentRegion()) {
-    return std::nullopt;
-  }
-  for (auto user : acc->getUsers()) {
-    if (auto load = dyn_cast<ttng::TMEMLoadOp>(user)) {
-      if (load->getParentRegion() == mmaOp->getParentRegion()) {
-        return std::make_pair(acc, load);
-      }
-    }
-  }
-  return std::nullopt;
-}
-
-ttng::TMEMAllocOp ttng::createTMemAlloc(OpBuilder &builder,
-                                        ttng::TMEMAllocOp oldTMemAllocOp,
-                                        bool multiBufferred, int numStages) {
-  Location loc = oldTMemAllocOp.getLoc();
-  auto oldRetType = oldTMemAllocOp.getType();
-  SmallVector<int64_t> shape = {oldRetType.getShape().begin(),
-                                oldRetType.getShape().end()};
-  if (multiBufferred) {
-    shape.insert(shape.begin(), numStages);
-  }
-  Type accMemDescType = triton::gpu::MemDescType::get(
-      shape, oldRetType.getElementType(), oldRetType.getEncoding(),
-      oldRetType.getMemorySpace(), /*mutableMemory=*/true);
-  return builder.create<ttng::TMEMAllocOp>(oldTMemAllocOp.getLoc(),
-                                           accMemDescType, nullptr);
-}
+using MMAInfo = ttng::MMAInfo;
 
 namespace {
 
@@ -60,11 +25,10 @@ const char *kPipelineStageAttrName = "triton.pipeline_stage";
 const char *kPipelineAttrName = "triton.pipeline";
 
 // Utils:
-void replaceAllUsesDominatedBy(Operation *domOp, Value newValue,
-                               Value oldValue) {
-  DominanceInfo domOpInfo(domOp->getParentOp());
+void replaceAllUsesDominatedBy(Operation *domOp, Value newValue, Value oldValue,
+                               DominanceInfo &domInfo) {
   oldValue.replaceUsesWithIf(newValue, [&](OpOperand &use) {
-    return domOpInfo.properlyDominates(domOp, use.getOwner());
+    return domInfo.properlyDominates(domOp, use.getOwner());
   });
 }
 
@@ -75,200 +39,6 @@ void annotateWithPipelineStage(IRRewriter &builder, Operation *op, int stage) {
 
 int getPipelineStage(Operation *op) {
   return op->getAttrOfType<IntegerAttr>(kPipelineStageAttrName).getInt();
-}
-
-struct MMAInfo {
-  struct AccOverridePoint {
-    Operation *op;
-    Value condition = nullptr;
-    Value initValue = nullptr;
-    int distance = 0;
-    bool isFlag = false;
-  };
-
-  ttng::TMEMAllocOp accAlloc; // Directly precedes the dot, allocating tmem
-                              // for the accumulator
-  ttng::TMEMLoadOp
-      accLoad; // Directly follows the dot, loading accumulator from tmem
-  std::optional<AccOverridePoint> accDef;
-  std::optional<int> yieldArgNo;
-  bool accIsMultiBuffered;
-
-  Value phase = nullptr;
-  Value barrierIdx = nullptr;
-  Value accInsertIdx = nullptr;
-  Value accExtractIdx = nullptr;
-  Value barrierAlloc = nullptr;
-};
-
-// Check if the accumulator is being used by the same MMA in the next iteration.
-// If so, return the yield argument number that the accumulator is being used
-// as. Also, check if accumulator has runtime divergent uses - uses that may not
-// be known at the compile time.
-std::optional<int> trackAccChain(scf::ForOp forOp, ttng::TMEMLoadOp accDef,
-                                 ttng::TMEMAllocOp accAlloc,
-                                 bool &hasDivergentUses) {
-  hasDivergentUses = false;
-  struct UseInfo {
-    Value value = nullptr;
-    std::optional<int> yieldArgNo = std::nullopt;
-    bool divergentUse = false;
-  };
-  SmallVector<UseInfo> queue;
-  std::optional<int> yieldArgNo = std::nullopt;
-  queue.push_back({accDef.getResult(), std::nullopt, false});
-  while (!queue.empty()) {
-    UseInfo info = queue.pop_back_val();
-    for (auto &use : info.value.getUses()) {
-      if (auto yieldOp = dyn_cast<scf::YieldOp>(use.getOwner())) {
-        if (yieldOp->getParentOp() == forOp) {
-          queue.push_back({forOp.getRegionIterArg(use.getOperandNumber()),
-                           use.getOperandNumber(), true}); // divergent use
-          continue;
-        }
-        if (auto ifOp = dyn_cast<scf::IfOp>(yieldOp->getParentOp())) {
-          queue.push_back({ifOp.getResult(use.getOperandNumber()),
-                           info.yieldArgNo, true}); // divergent use
-          continue;
-        }
-        assert(0 && "Unexpected use of accumulator");
-      } else if (auto selectOp = dyn_cast<arith::SelectOp>(use.getOwner())) {
-        queue.push_back({selectOp.getResult(), info.yieldArgNo, true});
-      } else if (use.getOwner() == accAlloc) {
-        yieldArgNo = info.yieldArgNo;
-      } else {
-        // Op other than yield or accAlloc. Mark as divergent use if
-        // we had to go through selectOp or ifOp.
-        hasDivergentUses = info.divergentUse;
-      }
-    }
-  }
-  return yieldArgNo;
-}
-
-SmallVector<Operation *> getDirectAccUses(ttng::TMEMLoadOp accDef) {
-  SmallVector<Operation *> accUses;
-  for (auto user : accDef.getResult().getUsers()) {
-    if (!isa<arith::SelectOp>(user) && !isa<scf::YieldOp>(user)) {
-      accUses.push_back(user);
-    }
-  }
-  return accUses;
-}
-
-std::optional<MMAInfo::AccOverridePoint>
-getAccOverridePointInLoop(scf::ForOp forOp, ttng::TMEMAllocOp accUse,
-                          ttng::TMEMLoadOp accDef) {
-  MMAInfo::AccOverridePoint accOverridePoint;
-  accOverridePoint.isFlag = false;
-  DenseSet<Value> seen;
-  Value v = accUse.getSrc();
-  if (v == nullptr) {
-    // Uninitialized accumulator means unused accumulator
-    accOverridePoint.op = accUse;
-    return accOverridePoint;
-  }
-  int dist = 0;
-  while (auto blockArg = dyn_cast<BlockArgument>(v)) {
-    if (!seen.insert(v).second) {
-      return std::nullopt;
-    }
-    assert(blockArg.getOwner() == forOp.getBody());
-    auto yieldOp = cast<scf::YieldOp>(blockArg.getOwner()->getTerminator());
-    v = yieldOp.getOperand(blockArg.getArgNumber() - 1);
-    dist++;
-  }
-  if (!v.getDefiningOp()) {
-    return std::nullopt;
-  }
-  accOverridePoint.distance = dist;
-  bool thenOverrides = false;
-  if (auto selectOp = dyn_cast<arith::SelectOp>(v.getDefiningOp())) {
-    accOverridePoint.op = selectOp;
-    bool trueIsConst =
-        (selectOp.getTrueValue().getDefiningOp<arith::ConstantOp>() != nullptr);
-    bool falseIsConst =
-        (selectOp.getFalseValue().getDefiningOp<arith::ConstantOp>() !=
-         nullptr);
-    if (trueIsConst && falseIsConst) {
-      // Both values are constant, so the select overrides unconditionally
-      accOverridePoint.initValue = v;
-      return accOverridePoint;
-    } else if (trueIsConst) {
-      accOverridePoint.initValue = selectOp.getTrueValue();
-      thenOverrides = true;
-    } else if (falseIsConst) {
-      accOverridePoint.initValue = selectOp.getFalseValue();
-      thenOverrides = false;
-    } else {
-      return std::nullopt;
-    }
-    accOverridePoint.condition = selectOp.getCondition();
-    if (!thenOverrides) {
-      IRRewriter builder(selectOp);
-      Value vTrue = builder.create<arith::ConstantOp>(
-          selectOp.getLoc(), builder.getBoolAttr(true));
-      accOverridePoint.condition = builder.create<arith::XOrIOp>(
-          selectOp.getLoc(), accOverridePoint.condition, vTrue);
-    }
-  } else if (v.getDefiningOp() != accDef) {
-    assert(!isa<scf::IfOp>(v.getDefiningOp()) &&
-           "Expected unconditional override op");
-    accOverridePoint.op = v.getDefiningOp();
-    accOverridePoint.initValue = v;
-  } else {
-    return std::nullopt;
-  }
-
-  return accOverridePoint;
-}
-
-std::optional<MMAInfo::AccOverridePoint>
-getAccUseFlagFalseInLoop(scf::ForOp forOp, Value useAccFlagUse) {
-  DenseSet<Value> seen;
-  Value v = useAccFlagUse;
-  int dist = 0;
-  while (auto blockArg = dyn_cast<BlockArgument>(v)) {
-    if (!seen.insert(v).second) {
-      return {};
-    }
-    assert(blockArg.getOwner() == forOp.getBody());
-    auto yieldOp = cast<scf::YieldOp>(blockArg.getOwner()->getTerminator());
-    v = yieldOp.getOperand(blockArg.getArgNumber() - 1);
-    dist++;
-  }
-  if (!v.getDefiningOp() || !forOp->isAncestor(v.getDefiningOp())) {
-    return std::nullopt;
-  }
-  assert(v.getType().isInteger(1));
-
-  IRRewriter builder(v.getDefiningOp()->getNextNode());
-  MMAInfo::AccOverridePoint accOverridePoint;
-  accOverridePoint.isFlag = true;
-  accOverridePoint.distance = dist;
-  Location loc = v.getDefiningOp()->getLoc();
-  auto vTrue =
-      builder.create<arith::ConstantOp>(loc, builder.getBoolAttr(true));
-  accOverridePoint.op = v.getDefiningOp();
-  accOverridePoint.condition = builder.create<arith::XOrIOp>(loc, v, vTrue);
-
-  return accOverridePoint;
-}
-
-std::optional<MMAInfo::AccOverridePoint>
-getAccOverrideOrFlagFalseInLoop(scf::ForOp forOp,
-                                ttng::MMAv5OpInterface mmaOp) {
-  auto tmemAllocAndLoad = getTMemAllocAndLoad(mmaOp);
-  assert(tmemAllocAndLoad.has_value() && "Expected tmem alloc and load");
-  auto [accAlloc, accLoad] = tmemAllocAndLoad.value();
-  auto accOverridePoint = getAccOverridePointInLoop(forOp, accAlloc, accLoad);
-
-  if (!accOverridePoint.has_value()) {
-    auto useAccFlag = mmaOp.useAccumulator();
-    accOverridePoint = getAccUseFlagFalseInLoop(forOp, useAccFlag);
-  }
-
-  return accOverridePoint;
 }
 
 void createInitStore(IRRewriter &builder, ttng::TMEMAllocOp allocOp,
@@ -282,39 +52,10 @@ void createInitStore(IRRewriter &builder, ttng::TMEMAllocOp allocOp,
                                     vTrue);
 }
 
-Operation *findNearestCommonDominator(ArrayRef<Operation *> ops,
-                                      DominanceInfo &domInfo) {
-  if (ops.size() == 0) {
-    return nullptr;
-  }
-  if (ops.size() == 1) {
-    return ops[0];
-  }
-  llvm::SmallPtrSet<Block *, 16> blocks;
-  for (auto op : ops) {
-    blocks.insert(op->getBlock());
-  }
-  Block *domBlock = domInfo.findNearestCommonDominator(blocks);
-  if (domBlock == nullptr) {
-    return nullptr;
-  }
-  SmallVector<Operation *> ancestorOps;
-  for (auto op : ops) {
-    ancestorOps.push_back(domBlock->findAncestorOpInBlock(*op));
-  }
-  Operation *dom = ancestorOps[0];
-  for (unsigned i = 1; i < ops.size(); i++) {
-    if (ancestorOps[i]->isBeforeInBlock(dom)) {
-      dom = ancestorOps[i];
-    }
-  }
-  return dom;
-}
-
 void updateAccUsesInLoop(IRRewriter &builder, scf::ForOp forOp, MMAInfo &info,
                          ttng::TMEMAllocOp newAlloc, int numStages) {
   DominanceInfo domInfo(forOp);
-  SmallVector<Operation *> directUses = getDirectAccUses(info.accLoad);
+  SmallVector<Operation *> directUses = ttng::getDirectAccUses(info.accLoad);
   if (!directUses.empty()) {
     Operation *domOp = findNearestCommonDominator(directUses, domInfo);
     assert(domOp != nullptr && "Could not find a common dominator");
@@ -361,7 +102,8 @@ void updateAccUsesOutsideLoop(IRRewriter &builder, scf::ForOp forOp,
 }
 
 void updateAccDefsInLoop(IRRewriter &builder, scf::ForOp forOp, MMAInfo &info,
-                         ttng::TMEMAllocOp newAlloc, int numStages) {
+                         ttng::TMEMAllocOp newAlloc, int numStages,
+                         DominanceInfo &domInfo) {
   assert(info.accDef.has_value());
   Operation *def = info.accDef->op;
   Value condition = info.accDef->condition;
@@ -382,12 +124,11 @@ void updateAccDefsInLoop(IRRewriter &builder, scf::ForOp forOp, MMAInfo &info,
   Value insWrap = builder.create<arith::CmpIOp>(loc, arith::CmpIPredicate::eq,
                                                 newInsertIdx, numStagesVal);
   newInsertIdx = builder.create<arith::SelectOp>(
-      loc, newInsertIdx.getType(), insWrap,
-      builder.create<arith::ConstantIntOp>(loc, 0, 32), newInsertIdx);
+      loc, insWrap, builder.create<arith::ConstantIntOp>(loc, 0, 32),
+      newInsertIdx);
   if (condition) {
-    newInsertIdx =
-        builder.create<arith::SelectOp>(loc, newInsertIdx.getType(), condition,
-                                        newInsertIdx, info.accInsertIdx);
+    newInsertIdx = builder.create<arith::SelectOp>(loc, condition, newInsertIdx,
+                                                   info.accInsertIdx);
   }
   annotateWithPipelineStage(builder, newInsertIdx.getDefiningOp(), 0);
 
@@ -397,12 +138,11 @@ void updateAccDefsInLoop(IRRewriter &builder, scf::ForOp forOp, MMAInfo &info,
   auto extWrap = builder.create<arith::CmpIOp>(loc, arith::CmpIPredicate::eq,
                                                newExtractIdx, numStagesVal);
   newExtractIdx = builder.create<arith::SelectOp>(
-      loc, newExtractIdx.getType(), extWrap,
-      builder.create<arith::ConstantIntOp>(loc, 0, 32), newExtractIdx);
+      loc, extWrap, builder.create<arith::ConstantIntOp>(loc, 0, 32),
+      newExtractIdx);
   if (info.accDef->condition) {
     newExtractIdx = builder.create<arith::SelectOp>(
-        loc, newExtractIdx.getType(), info.accDef->condition, newExtractIdx,
-        info.accExtractIdx);
+        loc, info.accDef->condition, newExtractIdx, info.accExtractIdx);
   }
   annotateWithPipelineStage(builder, newExtractIdx.getDefiningOp(), 1);
 
@@ -425,9 +165,9 @@ void updateAccDefsInLoop(IRRewriter &builder, scf::ForOp forOp, MMAInfo &info,
   // loop iteration)
   if (info.accDef->distance == 0) {
     replaceAllUsesDominatedBy(newInsertIdx.getDefiningOp(), newInsertIdx,
-                              info.accInsertIdx);
+                              info.accInsertIdx, domInfo);
     replaceAllUsesDominatedBy(newExtractIdx.getDefiningOp(), newExtractIdx,
-                              info.accExtractIdx);
+                              info.accExtractIdx, domInfo);
   }
 
   if (info.accDef->initValue && condition) {
@@ -444,7 +184,7 @@ void updateAccDefsInLoop(IRRewriter &builder, scf::ForOp forOp, MMAInfo &info,
 // tmem allocs.
 void hoistAndUseTMemAlloc(IRRewriter &builder, scf::ForOp forOp,
                           ttng::MMAv5OpInterface mmaOp, MMAInfo &info,
-                          int numStages) {
+                          int numStages, DominanceInfo &domInfo) {
   builder.setInsertionPoint(forOp);
   Value zero = builder.create<arith::ConstantIntOp>(forOp.getLoc(), 0, 32);
   Value one = builder.create<arith::ConstantIntOp>(forOp.getLoc(), 1, 32);
@@ -485,7 +225,7 @@ void hoistAndUseTMemAlloc(IRRewriter &builder, scf::ForOp forOp,
   }
 
   if (info.accIsMultiBuffered) {
-    updateAccDefsInLoop(builder, forOp, info, newAlloc, numStages);
+    updateAccDefsInLoop(builder, forOp, info, newAlloc, numStages, domInfo);
   }
 
   info.accLoad.erase();
@@ -496,7 +236,7 @@ void hoistAndUseTMemAlloc(IRRewriter &builder, scf::ForOp forOp,
 // Create multi-buffered barrier allocs and lower the MMA to MMA + wait barrier
 void createBarrierAndWaitOps(IRRewriter &builder, scf::ForOp forOp,
                              ttng::MMAv5OpInterface mmaOp, MMAInfo &info,
-                             int numStages) {
+                             int numStages, DominanceInfo &domInfo) {
   builder.setInsertionPoint(forOp);
   Value zero = builder.create<arith::ConstantIntOp>(forOp.getLoc(), 0, 32);
   Value one = builder.create<arith::ConstantIntOp>(forOp.getLoc(), 1, 32);
@@ -524,18 +264,19 @@ void createBarrierAndWaitOps(IRRewriter &builder, scf::ForOp forOp,
 
   // New barrierIdx and phase are in the first stage, so they can be used by
   // the ops that are ahead of them in either order or stages.
-  newBarrierIdx = builder.create<arith::SelectOp>(loc, newBarrierIdx.getType(),
-                                                  barWrap, zero, newBarrierIdx);
+  newBarrierIdx =
+      builder.create<arith::SelectOp>(loc, barWrap, zero, newBarrierIdx);
   replaceAllUsesDominatedBy(newBarrierIdx.getDefiningOp(), newBarrierIdx,
-                            info.barrierIdx);
+                            info.barrierIdx, domInfo);
   info.barrierIdx = newBarrierIdx;
   annotateWithPipelineStage(builder, info.barrierIdx.getDefiningOp(), 0);
 
   Value originalPhase = info.phase;
   Value newPhase = builder.create<arith::SelectOp>(
-      loc, info.phase.getType(), barWrap,
-      builder.create<arith::XOrIOp>(loc, info.phase, one), info.phase);
-  replaceAllUsesDominatedBy(newPhase.getDefiningOp(), newPhase, info.phase);
+      loc, barWrap, builder.create<arith::XOrIOp>(loc, info.phase, one),
+      info.phase);
+  replaceAllUsesDominatedBy(newPhase.getDefiningOp(), newPhase, info.phase,
+                            domInfo);
   info.phase = newPhase;
   annotateWithPipelineStage(builder, info.phase.getDefiningOp(), 0);
 
@@ -633,6 +374,7 @@ FailureOr<scf::ForOp> preProcessLoopForTC05MMAPipelining(scf::ForOp forOp,
   }
 
   IRRewriter builder(forOp->getContext());
+  DominanceInfo domInfo(forOp);
   for (auto op : mmaOps) {
     // Avoid pipelining if in the backward slice of the mmaOp there is an
     // operation that is already assigned a stage, as it would make the pipeline
@@ -648,69 +390,10 @@ FailureOr<scf::ForOp> preProcessLoopForTC05MMAPipelining(scf::ForOp forOp,
       continue;
     }
 
-    auto allocAndLoadOpt = getTMemAllocAndLoad(mmaOp);
-    if (!allocAndLoadOpt) {
+    std::optional<MMAInfo> mmaInfoOr = ttng::getMMAInfo(forOp, mmaOp, domInfo);
+    if (!mmaInfoOr)
       continue;
-    }
-    auto [accAlloc, accLoad] = allocAndLoadOpt.value();
-    bool hasDivergentUses = false;
-    std::optional<int> yieldArgNo =
-        trackAccChain(forOp, accLoad, accAlloc, hasDivergentUses);
-    if (hasDivergentUses) {
-      // If we can't tell for sure that the value is coming from the mma
-      // accumulator, skip.
-      continue;
-    }
-    if (yieldArgNo.has_value()) {
-      int accInitArgNo =
-          cast<BlockArgument>(accAlloc.getSrc()).getArgNumber() - 1;
-      assert(yieldArgNo.value() == accInitArgNo);
-    }
-
-    std::optional<MMAInfo::AccOverridePoint> accOverridePoint =
-        getAccOverrideOrFlagFalseInLoop(forOp, mmaOp);
-
-    if (accOverridePoint.has_value() && accOverridePoint->distance > 1) {
-      // We only support an override up to 1 iteration back.
-      continue;
-    }
-
-    SmallVector<Operation *> accUses = getDirectAccUses(accLoad);
-    DominanceInfo domOpInfo(forOp);
-    Operation *newAccLoadInsertPoint =
-        findNearestCommonDominator(accUses, domOpInfo);
-    // Check pipelining and multi-buffering constraints
-    // 1. Really needs multibuffering - if the acc is used unconditionally in
-    // the loop, or under different conditions. If we cannot multibuffer in this
-    // case, we may as well not pipeline at all, as we will have to wait after
-    // the dot in every loop iteration.
-    scf::IfOp topLevelIf =
-        newAccLoadInsertPoint
-            ? dyn_cast<scf::IfOp>(forOp.getBody()->findAncestorOpInBlock(
-                  *newAccLoadInsertPoint))
-            : nullptr;
-    bool requiresMultiBuffer = accUses.size() > 0 && !topLevelIf;
-    // If we override the acc in the loop, it is generally hard to handle it
-    // without multibuffering. We make an exception if it not a physical
-    // override of a value, but just setting a flag that acc is not used. In
-    // this case we don't need different buffer to store init value.
-    requiresMultiBuffer |=
-        accOverridePoint.has_value() && !accOverridePoint->isFlag;
-
-    // 2. If the acc is not owerwritten in the loop (by op other than the dot),
-    // it cannot be multi-buffered. This is because the overwrite is the only
-    // way to initialize next buffer without incurring a copy.
-    bool canMultiBuffer = accOverridePoint.has_value() &&
-                          !mlir::triton::getDisallowAccMultiBuffer(forOp);
-    if (requiresMultiBuffer && !canMultiBuffer) {
-      continue;
-    }
-
-    MMAInfo mmaInfo = {.accAlloc = accAlloc,
-                       .accLoad = accLoad,
-                       .accDef = accOverridePoint,
-                       .yieldArgNo = yieldArgNo,
-                       .accIsMultiBuffered = canMultiBuffer};
+    MMAInfo mmaInfo = std::move(*mmaInfoOr);
 
     builder.setInsertionPoint(forOp);
     Value zero = builder.create<arith::ConstantIntOp>(forOp.getLoc(), 0, 32);
@@ -743,8 +426,8 @@ FailureOr<scf::ForOp> preProcessLoopForTC05MMAPipelining(scf::ForOp forOp,
     appendToForOpYield(forOp, newYieldOperands);
 
     annotateWithPipelineStage(builder, mmaOp, 0);
-    hoistAndUseTMemAlloc(builder, forOp, mmaOp, mmaInfo, numStages);
-    createBarrierAndWaitOps(builder, forOp, mmaOp, mmaInfo, numStages);
+    hoistAndUseTMemAlloc(builder, forOp, mmaOp, mmaInfo, numStages, domInfo);
+    createBarrierAndWaitOps(builder, forOp, mmaOp, mmaInfo, numStages, domInfo);
   }
 
   return forOp;
